introduce modularization and change states to enum
